+++ src/main/php/util/log/SyslogAppender.class.php
@@ -7,15 +7,15 @@
  * @see  php://syslog
  */  
 class SyslogAppender extends Appender {
-  protected $lastIdentifier= false;
+  protected static $lastIdentifier= false;
   public $identifier, $facility;
 
   /**
    * Constructor
    *
-   * @param   string identifier default NULL if omitted, defaults to script's filename
-   * @param   int facility default LOG_USER
-   * @see     php://openlog for valid facility values
+   * @param  string $identifier default NULL if omitted, defaults to script's filename
+   * @param  int $facility default LOG_USER
+   * @see    php://openlog for valid facility values
    */
   public function __construct($identifier= null, $facility= LOG_USER) {
     $this->identifier= $identifier;
@@ -25,7 +25,7 @@ class SyslogAppender extends Appender {
   /**
    * Append data
    *
-   * @param   util.log.LoggingEvent event
+   * @param  util.log.LoggingEvent $event
    */ 
   public function append(LoggingEvent $event) {
     static $map= [
@@ -36,14 +36,14 @@ class SyslogAppender extends Appender {
       LogLevel::NONE    => LOG_NOTICE
     ];
 
-    if ($this->identifier != $this->lastIdentifier) {
+    if ($this->identifier !== self::$lastIdentifier) {
       closelog();
       openlog(
-        $this->identifier ? $this->identifier : basename($_SERVER['PHP_SELF']), 
+        $this->identifier ?: basename($_SERVER['PHP_SELF']), 
         LOG_ODELAY | LOG_PID, 
         $this->facility
       );
-      $this->lastIdentifier= $this->identifier;
+      self::$lastIdentifier= $this->identifier;
     }
 
     $l= $event->getLevel();
@@ -54,15 +54,15 @@ class SyslogAppender extends Appender {
    * Finalize this appender - is called when the logger shuts down
    * at the end of the request.
    *
+   * @return void
    */
   public function finalize() {
     closelog();
-    $this->lastIdentifier= false;
+    self::$lastIdentifier= false;
   }
 
   /**
-   * Destructor.
-   *
+   * Destructor
    */
   public function __destruct() {
     $this->finalize();