+++ ChangeLog.md
@@ -14,6 +14,12 @@ XP Framework Core ChangeLog
   [Composer](https://getcomposer.org/)
   (@thekid)
 
+### Bugfixes
+
+* Fixed util.log.SyslogAppender problems with multiple instances logging
+  to the same identifier and facility.
+  (@mikey179, @thekid)
+
 ## 6.4.2 / 2015-08-05
 
 ### Bugfixes